% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{(integer) number of trials}

\item{prob}{(numeric) probability of success}
}
\value{
a data.frame with two classes: c("bincum", "data.frame")
}
\description{
creates a data.frame with two classes: c("bincum", "data.frame")
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
