% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{(integer) number of trials}

\item{prob}{(numeric) probability of success}
}
\value{
the kurtosis of a binomial distribution
}
\description{
calculates the kurtosis of a binomial distribution
}
\examples{
bin_kurtosis(10, 0.3)
}
