% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean()}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{(integer) number of trials}

\item{prob}{(numeric) probability of success}
}
\value{
the mean of a binomial distribution
}
\description{
calculates the mean of a binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
