% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{(integer) number of successes}

\item{trials}{(integer) number of trials}

\item{prob}{(numeric) probability of success}
}
\value{
the probability of k successes on n repeated trials
}
\description{
calculates the probability of k successes on n repeated trials
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
}
