% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{(integer) number of trials}

\item{prob}{(numeric) probability of success}
}
\value{
the skewness of a binomial distribution
}
\description{
calculates the skewness of a binomial distribution
}
\examples{
bin_skewness(10, 0.3)
}
