% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{(integer) number of trials}

\item{prob}{(numeric) probability of success}
}
\value{
a binomial random variable object of class "binvar"
}
\description{
creates a binomial random variable object of class "binvar"
}
\examples{
bin_variable(trials = 10, p = 0.3)
}
