% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{(integer) number of trials}

\item{prob}{(numeric) probability of success}
}
\value{
the variance of a binomial distribution
}
\description{
calculates the variance of a binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
